Implement functionality for hex to binary conversion
This one was a little tedious to implement, but I was able to make
it more general and avoided having to hard-code all the cases, which
I am happy about. I still need to go in and test it more thoroughly
because I admittedly was lazy with testing, but I now have some base
functionality for all types of conversions.